pyproject.toml
[project]
name = "aitest"
version = "0.1.0"
description = "my ai test"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.26.0",
    "python-dotenv>=0.19.0",
    "pydantic>=2.0.0",
    "chromadb>=1.5.5",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "pytest>=9.0.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
aitest = "main:main"
